refactor: move app startup side effects out of App.tsx
