+++ frontend/src/App.tsx
@@ -1,40 +1,20 @@
-import { useEffect, useMemo } from "react";
+import { useMemo } from "react";
 import { Routes, Route } from "react-router-dom";
 import { CssBaseline, Box } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
-import {
-    getAggregatedDates,
-    getSites,
-    getStopPoints,
-    getTodayStopPointRoutes,
-} from "./store/actions";
-import { useAppDispatch, useAppSelector } from "./store/store";
+import { useAppSelector } from "./store/store";
 import { SidebarPresenter } from "./presenters/sidebarPresenter";
 import { AccountPresenter } from "./presenters/accountPresenter";
 import { LoginPresenter } from "./presenters/loginPresenter";
-import { initializeAuthSync } from "./store/authThunks";
 import { GlobalSnackbarPresenter } from "./presenters/globalSnackbarPresenter";
 import { ROUTES, type RouteConfig } from "./routes";
 import { getAppStylePreferenceCB } from "./store/selectors";
 import { createAppMuiTheme } from "./theme/muiTheme";
+import { AppStartupPresenter } from "./presenters/appStartupPresenter";
 
 function App() {
-    const dispatch = useAppDispatch();
     const appStyle = useAppSelector(getAppStylePreferenceCB);
     const muiTheme = useMemo(() => createAppMuiTheme(appStyle), [appStyle]);
-
-    // Fetch sites, stop points, and aggregated dates on app load
-    useEffect(() => {
-        dispatch(getSites());
-        dispatch(getStopPoints());
-        dispatch(getTodayStopPointRoutes());
-        dispatch(getAggregatedDates());
-
-        // return the unsubscribe function to stop the firebase connection
-        // if this component ever unmounts to prevent memory leaks.
-        const unsubscribeCB = dispatch(initializeAuthSync());
-        return unsubscribeCB;
-    }, [dispatch]);
 
     function renderRouteCB(route: RouteConfig) {
         return <Route key={route.path} path={route.path} element={route.element} />;
@@ -47,6 +27,7 @@ function App() {
                 className="relative h-screen w-screen overflow-hidden"
                 sx={{ bgcolor: "background.default", color: "text.primary" }}
             >
+                <AppStartupPresenter />
                 <SidebarPresenter />
                 <Routes>
                     {ROUTES.map(renderRouteCB)}
